+++ client/views/Teacher/QuestionCreate/View.jsx
@@ -1,24 +1,8 @@
 import React from 'react';
 
-TeacherQuestionCreateView = React.createClass({
-  mixins: [ReactMeteorData],
+import TeacherQuestionCreateForm from './Form/index.jsx';
 
-  getMeteorData() {
-    const userId = Meteor.userId();
-
-    const handles = {
-      subjects: Meteor.subscribe('PublicSubjects'),
-      tags: Meteor.subscribe('PublicTags'),
-    };
-
-    return {
-      user: Meteor.user(),
-      ready: _.mapValues(handles, h => h.ready()),
-      subjects: Fetch.Public().subjects().fetch(),
-      tags: Fetch.Public().tags().fetch(),
-    };
-  },
-
+const TeacherQuestionCreateView = React.createClass({
   render() {
 
     return (
@@ -33,3 +17,4 @@ TeacherQuestionCreateView = React.createClass({
     );
   },
 });
+export default TeacherQuestionCreateView;
